dinnertonight/requirements.txt
Django==5.1.3
transformers==4.46.2
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
python-dotenv
requests